Push to "origin" in tag_git_repos function
The repositories all use "origin" as the upstream repository name, so
this name can be used instead of passing the URL explicitly to
tag_git_repos function.
This simplifies the function by making it uniform and allows renaming
the repositories paths.

The "origin" remotes had to change to use SSH protocol to allow the push
of tags. The URL had to be added to the build-release-notes command and
modified for builds repository.